+++ src/StkTable/StkTable.vue
@@ -376,7 +376,7 @@ const props = withDefaults(
          * "v" - 仅展示竖线
          * "body-v" - 仅表体展示竖线
          */
-        bordered?: boolean | 'h' | 'v' | 'body-v';
+        bordered?: boolean | 'h' | 'v' | 'body-v' | 'body-h';
         /**
          * 自动重新计算虚拟滚动高度宽度。默认true
          * [非响应式]