+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 0.8.12
 * Bugfix
   - fix: customCell -> `CustomCellProps['expanded']` always return undefined
+  - fix: type CustomCellProps['expanded'] and CustomCellProps['treeExpanded'] to be optional.
+  - fix:  If set props['headerDrag'] an object, it will cause warning, which is validated by vue. Same as `autoRowHeight` , `rowActive`, `colResizable`.
 
 ## 0.8.11
 * Change